Move `codemirror.ts` to `@/components/CodeEditor/`
